requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
aws-auth-helper==1.5.5
billiard==3.3.0.23
boto3==1.12.20
botocore==1.15.20
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj==0.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
django-filter==2.2.0
django-filters==0.2.1
django-heroku==0.3.1
django-import-export==2.2.0
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.0.1
filelock==3.0.12
gitignore==0.0.8
gunicorn==19.7.1
heroku==0.1.4
idna==2.9
importlib-metadata==1.5.0
jdcal==1.4.1
jmespath==0.9.5
kombu==3.0.37
MarkupPy==1.14
more-itertools==8.2.0
mysql==0.0.2
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1

oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.3
Pillow==5.4.1
pipenv==2018.11.26
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyparsing==2.4.7
pystache==0.5.4
pytest==5.4.1
python-dateutil==2.8.1
pytz==2018.9
PyYAML==5.2
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
static==1.1.1
static3==0.7.0
tablib==2.0.0
text-unidecode==1.2
toml==0.10.0
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.10
virtualenv-clone==0.5.3
wcwidth==0.1.9
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
